Upgrade

+++ ManagerAPI.Shared/DTOs/WM/WorkingWeekStatDto.cs
@@ -7,11 +7,34 @@ namespace ManagerAPI.Shared.DTOs.WM
     /// </summary>
     public class WorkingWeekStatDto
     {
-        public int HourSum { get; set; }
+        /// <summary>
+        /// Hour Sum
+        /// </summary>
+        public decimal HourSum { get; set; }
+
+        /// <summary>
+        /// Hour Avg
+        /// </summary>
         public double HourAvg { get; set; }
+
+        /// <summary>
+        /// Active Days
+        /// </summary>
         public int ActiveDays { get; set; }
+
+        /// <summary>
+        /// Work days
+        /// </summary>
         public int WorkDays { get; set; }
+
+        /// <summary>
+        /// Counts
+        /// </summary>
         public List<WorkingDayTypeCountDto> Counts { get; set; }
+
+        /// <summary>
+        /// Fields
+        /// </summary>
         public List<WorkingFieldListDto> Fields { get; set; }
     }
 }
